backend/requirements.txt
# KrishiMitra AI — Cleaned Requirements (no duplicates)
# Generated by Antigravity Fix Script

# ── Core Framework ──────────────────────────
Django>=5.2,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# ── Database ─────────────────────────────────
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# ── Production Server ─────────────────────────
gunicorn>=21.0.0
whitenoise>=6.5.0

# ── Environment ───────────────────────────────
python-decouple>=3.8
django-environ>=0.11.0

# ── AI/ML ────────────────────────────────────
google-generativeai>=0.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# ── HTTP & APIs ───────────────────────────────
requests>=2.31.0
urllib3<2.0.0
aiohttp>=3.8.0

# ── Caching & Tasks ──────────────────────────
redis>=4.5.0
celery>=5.3.0
django-redis>=5.4.0
django-celery-beat>=2.5.0

# ── Security ─────────────────────────────────
djangorestframework-simplejwt>=5.3.0
django-ratelimit>=4.0.0
cryptography>=41.0.0

# ── Geolocation ──────────────────────────────
geopy>=2.3.0

# ── Text-to-Speech ────────────────────────────
gTTS>=2.3.0

# ── API Documentation ─────────────────────────
drf-spectacular>=0.26.0

# ── Monitoring ───────────────────────────────
sentry-sdk>=1.32.0
psutil>=5.9.0

# ── Data Processing ──────────────────────────
python-dateutil>=2.8.0
pytz>=2023.3
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ── Development ──────────────────────────────
django-debug-toolbar>=4.1.0
python-dotenv>=1.0.0